Refactor timestamps (#1016)
In essence, all datetime strings of the REST API are serialized as ISO 8601 (aka JSON time) strings incl. time offsets. This also holds for the "min/max takedate" of the albums which has previously been sent as a localized string including the year and month only. Moreover, for the take date of photos the original timezone is reported to the client, too.

* REST API returns JSON time format
* Re-compiled front end
* Fixed to erronous strtotime()
* Added a patch to Laravel's base model, added custon cast, accessor and mutator
* Fixed Photo model: Mutators and Casts must not be used together
* Fixed tests
* Fixed migration and re-compiled frontend
* Dumped new minor version
* More work on comment in Extractor
* Fixed missing upgrade for album sorting in migration
* Merge of Lychee-front
* Switched Lychee frontend to master