Update HasRandomIDAndLegacyTimeBasedID.php
Modification of ID generation based on recommendation by associate who is a PHP developer. This will avoid having a hyphen at the end of URLs which seems to cause issues with certain apps/devices that do URL parsing.

In an exceptionally rare, but not impossible situation, an ID of "--------------" would result in "".