No middleware necessary + modal instead of alert
The try catch block in axios got me thinking. The cookie is set to
expire with the session. As a result we can track this to know whether
it has expired or not.

Trigger a specific event which is capture by the Error page and display
a modal instead of a default JS alert. This is UI and UX friendlier.